feat: refactor cmd
