+++ .goreleaser.yaml
@@ -1,10 +1,6 @@
-# Make sure to check the documentation at https://goreleaser.com
 before:
   hooks:
-    # You may remove this if you don't use go modules.
     - go mod tidy
-    # you may remove this if you don't need go generate
-    - go generate ./...
 builds:
   - env:
       - CGO_ENABLED=0
@@ -12,7 +8,8 @@ builds:
       - linux
       - windows
       - darwin
-    binary: ctp
+    main: "./cmd/ctp"
+    binary: "ctp"
 archives:
   - replacements:
       darwin: Darwin
