+++ cmd/ctp/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/catppuccin/cli/internal/cli"
+
+func main() {
+	cli.Execute()
+}
