+++ internal/cli/new.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"github.com/catppuccin/cli/internal/ui"
