+++ internal/cli/root.go
@@ -1,7 +1,7 @@
 /*
 Copyright © 2022 NAME HERE <EMAIL ADDRESS>
 */
-package cmd
+package cli
 
 import (
 	"os"
