--- main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/catppuccin/cli/cmd"
-
-func main() {
-	cmd.Execute()
-}
